% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkaholic.R
\encoding{UTF-8}
\name{extract_polkaholic_events}
\alias{extract_polkaholic_events}
\title{extract_polkaholic_events}
\usage{
extract_polkaholic_events(dat)
}
\arguments{
\item{dat}{data.table output from `get_polkaholic_events`}
}
\value{
data.table
}
\description{
Get events from the Polkadot blockchain from the Polkaholic api
https://docs.polkaholic.io
}
\examples{
t1 <- get_polkaholic_events("karura", module = "dex", call = 'Swap',
startBlock=1, endBlock=2208550, nobs=10)
t2 <- extract_polkaholic_events(t1)

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Extract events from the Polkadot blockchain from the Polkaholic api}
