% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkaholic.R
\encoding{UTF-8}
\name{get_polkaholic_chainlog}
\alias{get_polkaholic_chainlog}
\title{get_polkaholic_chainlog}
\usage{
get_polkaholic_chainlog(chain)
}
\arguments{
\item{chain}{string or integer specifying the target parachain}
}
\value{
data.table
}
\description{
Get events from the Polkadot blockchain from the Polkaholic api
https://docs.polkaholic.io
}
\examples{
tmp <- get_polkaholic_chainlog("astar")

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get events from the Polkadot blockchain from the Polkaholic api}
