% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkaholic.R
\encoding{UTF-8}
\name{get_polkaholic_chains}
\alias{get_polkaholic_chains}
\title{get_polkaholic_chains}
\usage{
get_polkaholic_chains()
}
\arguments{
\item{None}{}
}
\value{
data.table
}
\description{
Get supported chains from the Polkaholic api
https://docs.polkaholic.io
}
\examples{
get_polkaholic_chains()

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get events from the Polkadot blockchain from the Polkaholic api}
