% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkaholic.R
\encoding{UTF-8}
\name{get_polkaholic_events}
\alias{get_polkaholic_events}
\title{get_polkaholic_events}
\usage{
get_polkaholic_events(
  chain,
  module,
  call,
  startDate,
  endDate,
  startBlock = NA,
  endBlock = NA,
  nobs = 1000
)
}
\arguments{
\item{chain}{string or integer specifying the target parachain}

\item{module}{string indicating which Polkadot module to use; leave blank for all.}

\item{call}{string indicating the module call (i.e. "Swap"); leave blank for all.}

\item{startDate}{string indicating the end date; leave blank for all.}

\item{endDate}{string indicating the start date; leave blank for all.}
}
\value{
data.table
}
\description{
Get events from the Polkadot blockchain from the Polkaholic api
https://docs.polkaholic.io
}
\examples{
get_polkaholic_events("karura", module = "dex", call = 'Swap',
startDate='2022-07-01', endDate='2022-07-02', nobs = 5)

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get events from the Polkadot blockchain from the Polkaholic api}
