% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkaholic.R
\encoding{UTF-8}
\name{get_polkaholic_hash}
\alias{get_polkaholic_hash}
\title{get_polkaholic_hash}
\usage{
get_polkaholic_hash(TxHash)
}
\arguments{
\item{TxHash}{string hash to lookup}
}
\value{
data.table
}
\description{
Get supported chains from the Polkaholic api
https://docs.polkaholic.io
}
\examples{
get_polkaholic_hash(TxHash = '0x044e63a8548c51bac4f3cdf71cc3bc31e63b460f7652030604aee4361f1b1878')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get hash lookup from the Polkadot blockchain from the Polkaholic api}
