% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polkaholic.R
\encoding{UTF-8}
\name{get_polkaholic_transaction}
\alias{get_polkaholic_transaction}
\title{get_polkaholic_transaction}
\usage{
get_polkaholic_transaction(TxHash)
}
\arguments{
\item{TxHash}{string transaction hash}
}
\value{
data.table
}
\description{
Get supported chains from the Polkaholic api
https://docs.polkaholic.io
}
\examples{
get_polkaholic_transaction(TxHash = '0x8b91d038421d5aba4b3a651d70923fb0d41423b914d6dad910637aa2c9a2ad70')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get events from the Polkadot blockchain from the Polkaholic api}
