% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_account_token}
\alias{get_subscan_account_token}
\alias{get_subscan_account_tokens}
\title{get_subscan_account_token}
\usage{
get_subscan_account_tokens(network = "KaruraA", addr)
}
\arguments{
\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}
}
\value{
list
}
\description{
Get tokens from the Subscan api
https://docs.api.subscan.io
}
\examples{
get_subscan_account_tokens('Acala', addr =
'23M5ttkmR6Kco5p3LFGKMpMv4zvLkKdUQWW1wGGoV8zDX3am')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get tokens for a specific from the Subscan api}
