% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_accounts}
\alias{get_subscan_accounts}
\title{get_subscan_accounts}
\usage{
get_subscan_accounts(network = "Karura", nobs = 100, start_page = 1)
}
\arguments{
\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}
}
\value{
list
}
\description{
Get accounts from the Subscan api
https://docs.api.subscan.io
}
\examples{
get_subscan_accounts()
get_subscan_accounts(network = 'Acala')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get accounts from the Subscan api}
