% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_events}
\alias{get_subscan_events}
\title{get_subscan_events}
\usage{
get_subscan_events(nobs = 10, network = "Karura", ss_module = "", ss_call = "")
}
\arguments{
\item{nobs}{integer how many transactions to pull}

\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}

\item{ss_module}{string indicating which module to pull; leave blank for all.}

\item{ss_call}{string indicating which function call to pull; leave blank for all.}
}
\value{
data.table
}
\description{
Get events from the Polkadot blockchain from the Subscan api
https://docs.api.subscan.io
}
\examples{
tmp <- get_subscan_events(nobs = 111); dim(tmp$core_data)
tmp <- get_subscan_events(nobs = 10, network = 'Karura', ss_module = 'dex', ss_call = 'Swap')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get events from the Polkadot blockchain from the Subscan api}
