% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_price_converter}
\alias{get_subscan_price_converter}
\title{get_subscan_price_converter}
\usage{
get_subscan_price_converter(
  network = "Karuka",
  time = 957105,
  value = 1000,
  from = "USD",
  quote = "DOT"
)
}
\arguments{
\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}

\item{start}{date start time.}

\item{end}{date end time.}
}
\value{
list
}
\description{
Get the amount of the target currency base on the amount of source currency given historical price from the Subscan api
https://docs.api.subscan.io
}
\examples{
get_subscan_price_converter('Polkadot', time = 957105,
value = 1000, from = 'USD', quote = 'DOT')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get the amount of the target currency base on the amount of source currency given historical price from the Subscan api}
