% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_price_history}
\alias{get_subscan_price_history}
\title{get_subscan_price_history}
\usage{
get_subscan_price_history(
  network = "Karura",
  start = "2021-11-01",
  end = "2021-11-02"
)
}
\arguments{
\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}

\item{start}{date start time.}

\item{end}{date end time.}
}
\value{
list
}
\description{
Get average price for a date range from the Subscan api
https://docs.api.subscan.io
}
\examples{
get_subscan_price_history(network = 'Darwinia', start = '2021-11-01', end = '2021-11-02')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get average price for a date range from the Subscan api}
