% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_token}
\alias{get_subscan_token}
\title{get_subscan_token}
\usage{
get_subscan_token(network = "Karura")
}
\arguments{
\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}
}
\value{
list
}
\description{
Get tokens from the Subscan api
https://docs.api.subscan.io
}
\examples{
get_subscan_token()
get_subscan_token(network = 'acala')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get tokens from the Subscan api}
