% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subql_query.R
\docType{data}
\name{stagingStr}
\alias{stagingStr}
\title{library(jsonlite)
library(data.table)
library(ghql)
x <- GraphqlClient$new()}
\format{
An object of class \code{character} of length 1.
}
\usage{
stagingStr
}
\description{
#' @author Roger J. Bos, \email{roger.bos@gmail.com}
#' @export
mysort <- function(a, b) {
  ifelse(a < b, a %+% ":" %+% b, b %+% ":" %+% a)
}
}
\keyword{datasets}
