% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\name{`\%+\%`}
\alias{`\%+\%`}
\alias{\%+\%}
\title{Helper function wrapper to paste0()}
\usage{
a \%+\% b
}
\description{
Helper function wrapper to paste0()
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
