% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscan.R
\encoding{UTF-8}
\name{get_subscan_price}
\alias{get_subscan_price}
\alias{get_subscan_currencies}
\title{get_subscan_price}
\usage{
get_subscan_price(network = "Polkadot", time = 957105)

get_subscan_currencies(network = "Polkadot")
}
\arguments{
\item{network}{string indicating which Polkadot endpoint to use; defaults to 'Karura'.}

\item{time}{integer block time or unix time.}
}
\value{
list

list
}
\description{
Get price for a specific block time from the Subscan api
https://docs.api.subscan.io

Get list of currencies that support price query and conversion
https://docs.api.subscan.io
}
\examples{
get_subscan_price()

get_subscan_currencies('Polkadot')

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{Get list of currencies that support price query and conversion}
\concept{Get price for a specific block time from the Subscan api}
